+++ src/lib/theme/SideBar.svelte
@@ -9,6 +9,6 @@
 </script>
 
 <aside class="sidebar md:mr-0 md:mt-0 md:w-2/5 md:ml-6 block content-center content-va-on" >
-      <GoodReadsWidget goodReads={sidebarData.goodReads} />
-      <TagsCloudWidget tagCloud={sidebarData.tagCloud} />
+      <GoodReadsWidget goodReads={sidebarData?.goodReads} />
+      <TagsCloudWidget tagCloud={sidebarData?.tagCloud} />
 </aside>